+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [11.0.0] - 2018-12-11
+### Added
+- Export the `withForm` HOC for accessing form data from external components.
+
+### Removed
+- The `MultiImageField` component.
+
 ## [10.1.0] - 2018-12-11
 ### Added
 - Support for drag and drop on the `ImageField` component.